removed duplicated code and set up public functions for loading and saving data
